feat: Adds verify_skipped command line option
This forces a file verification and database insertion
if the file is correctly verified. This is useful if
the database got deleted and you already have downloaded
some images. In normal operation those images will be
skipped and not verified even if not in the database and
thus not inserted in this database. The option allows to
verify skipped image and, when verified, insert them in
the database.